+++ packages/engine-svg/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 2.0.0-beta.12 - 2023-07-08
+## 2.0.0-beta.12 - 2025-07-14
 
 ### Added